+++ diesel/src/mysql/query_builder/query_fragment_impls.rs
@@ -1,3 +1,4 @@
+use crate::expression::operators::Concat;
 use crate::mysql::Mysql;
 use crate::query_builder::insert_statement::DefaultValues;
 use crate::query_builder::locking_clause::{ForShare, ForUpdate, NoModifier, NoWait, SkipLocked};
@@ -41,6 +42,24 @@ impl QueryFragment<Mysql> for NoWait {
 impl QueryFragment<Mysql, crate::mysql::backend::MysqlStyleDefaultValueClause> for DefaultValues {
     fn walk_ast<'b>(&'b self, mut out: AstPass<'_, 'b, Mysql>) -> QueryResult<()> {
         out.push_sql("() VALUES ()");
+        Ok(())
+    }
+}
+
+impl<L, R> QueryFragment<Mysql, crate::mysql::backend::MysqlConcatClause> for Concat<L, R>
+where
+    L: QueryFragment<Mysql>,
+    R: QueryFragment<Mysql>,
+{
+    fn walk_ast<'b>(
+        &'b self,
+        mut out: crate::query_builder::AstPass<'_, 'b, Mysql>,
+    ) -> crate::result::QueryResult<()> {
+        out.push_sql("CONCAT(");
+        self.left.walk_ast(out.reborrow())?;
+        out.push_sql(",");
+        self.right.walk_ast(out.reborrow())?;
+        out.push_sql(")");
         Ok(())
     }
 }